[fix] id의 타입이 Long 이므로 findById 의 인자를 Long으로 수정

+++ src/main/java/kr/codesqaud/cafe/domain/article/repository/ArticleRepository.java
@@ -9,7 +9,7 @@ public interface ArticleRepository {
 
 	void save(Article article);
 
-	Optional<Article> findById(Integer id);
+	Optional<Article> findById(Long id);
 
 	List<Article> findAll();
